Fixes a race condition in IabHelper.
The IabHelper class has a dispose() method that does various cleanup tasks
(unbinds the ServiceConnection, nulls out instance fields, etc). This method
is intended to be called from the host activity's onDestroy method. IabHelper
also includes *Async methods that can perform billing operations on a back-
ground thread. However, this combination produces a race condition that can
cause runtime crashes. If dispose() is called while an async operation is
in progress, the async operation may crash, e.g. because mDisposed is false
(IllegalStateException) or mContext is null (NullPointerException).

The fix has two parts:

1) Have dispose() check upfront whether an async operation is in progress,
and throw an exception if so.

2) Add a disposeWhenFinished() method that checks for an in-progress async
operation and queues the dispose to occur after it completes.